move profile mutation to `viewer.`
