+++ server/routers/_app.ts
@@ -3,7 +3,6 @@
  */
 import { createRouter } from "../createRouter";
 import { bookingRouter } from "./booking";
-import { userRouter } from "./user";
 import { viewerRouter } from "./viewer";
 
 /**
@@ -24,7 +23,6 @@ export const appRouter = createRouter()
    */
   // .formatError(({ shape, error }) => { })
   .merge("viewer.", viewerRouter)
-  .merge("booking.", bookingRouter)
-  .merge("user.", userRouter);
+  .merge("booking.", bookingRouter);
 
 export type AppRouter = typeof appRouter;
