+++ server/routers/viewer.tsx
@@ -1,4 +1,15 @@
+import { TRPCError } from "@trpc/server";
+import { z } from "zod";
+
+import checkPremiumUsername from "@ee/lib/core/checkUsername";
+
+import checkRegularUsername from "@lib/core/checkUsername";
+import slugify from "@lib/slugify";
+
 import { createProtectedRouter } from "../createRouter";
+
+const checkUsername =
+  process.env.NEXT_PUBLIC_APP_URL === "https://cal.com" ? checkPremiumUsername : checkRegularUsername;
 
 // routes only available to authenticated users
 export const viewerRouter = createProtectedRouter()
@@ -60,5 +71,54 @@ export const viewerRouter = createProtectedRouter()
       });
 
       return bookings;
+    },
+  })
+  .mutation("updateProfile", {
+    input: z.object({
+      username: z.string().optional(),
+      name: z.string().optional(),
+      description: z.string().optional(),
+      avatar: z.string().optional(),
+      timeZone: z.string().optional(),
+      weekStart: z.string().optional(),
+      hideBranding: z.boolean().optional(),
+      theme: z.string().optional(),
+      completedOnboarding: z.boolean().optional(),
+      locale: z.string().optional(),
+    }),
+    async resolve({ input, ctx }) {
+      const { user, prisma } = ctx;
+      const { name, avatar, timeZone, weekStart, hideBranding, theme, completedOnboarding, locale } = input;
+      let username: string | undefined = undefined;
+
+      if (input.username) {
+        username = slugify(input.username);
+        // Only validate if we're changing usernames
+        if (username !== user.username) {
+          const response = await checkUsername(username);
+          if (response.status !== 200) {
+            const { message } = await response.json();
+            throw new TRPCError({ code: "BAD_REQUEST", message });
+          }
+        }
+      }
+
+      await prisma.user.update({
+        where: {
+          id: user.id,
+        },
+        data: {
+          username,
+          name,
+          avatar,
+          timeZone,
+          weekStart,
+          hideBranding,
+          theme,
+          completedOnboarding,
+          locale,
+          bio: input.description,
+        },
+      });
     },
   });
